Move dimension reduction into separate file.
